--- src/unicorn_eval/adaptors/segmentation/baseline_linear_upsample_conv3d/v1/__init__.py
@@ -1,6 +1,0 @@
-from unicorn_eval.adaptors.segmentation.baseline_linear_upsample_conv3d.v1.main import \
-    UnicornLinearUpsampleConv3D_V1
-
-__all__ = [
-    "UnicornLinearUpsampleConv3D_V1",
-]
