+++ src/unicorn_eval/adaptors/segmentation/baseline_segmentation_upsampling_3d/v1/main.py
@@ -12,8 +12,6 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-# TODO: refactor these adaptors in one per folder
-
 from __future__ import annotations
 
 import numpy as np
